saleh_dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
chromadb-client==0.5.7
python-multipart==0.0.9
